+++ src/modulon/event/sink/__init__.py
@@ -1,0 +1,9 @@
+from modulon.core.lookup import Lookup
+from modulon.event.sink.archive import AbstractEventArchiveFactory
+
+
+class EventSinkSubGraph(Lookup):
+    archive: AbstractEventArchiveFactory
+
+
+__all__ = ["EventSinkSubGraph"]