+++ src/modulon/event/source/__init__.py
@@ -1,0 +1,9 @@
+from modulon.core.lookup import Lookup
+from modulon.event.source.replay import AbstractEventReplayFactory
+
+
+class EventSourceSubGraph(Lookup):
+    replay: AbstractEventReplayFactory
+
+
+__all__ = ["EventSourceSubGraph"]